backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9

# Database
sqlmodel==0.0.21
psycopg[binary]==3.2.1
alembic==1.13.1

# Validation & Settings
pydantic==2.8.0
pydantic-settings==2.3.0
email-validator==2.2.0

# Auth
pyjwt==2.9.0
python-jose[cryptography]==3.3.0
pwdlib[argon2,bcrypt]==0.2.1

# HTTP & Utils
httpx==0.27.0
requests==2.32.3
emails==0.6
tenacity==8.5.0
jinja2==3.1.4
python-socketio==5.11.0

# Monitoring (optional)
sentry-sdk==1.40.6
